requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
faiss-cpu==1.7.2
filelock==3.16.0
fsspec==2024.9.0
huggingface-hub==0.24.6
idna==3.8
iniconfig==2.0.0
 Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==1.21.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyPDF2==3.0.1
pytest==8.3.3
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
